fips: Remove specific keystore handling
Revert fips specific behavior with NewGCMWithRandomNonce as this is
not implemented by the go/microsoft toolchain.
